tests: Nightly fixes (2025-08-20)
