+++ ci/test/pipeline.template.yml
@@ -384,7 +384,7 @@ steps:
     depends_on: build-x86_64
     timeout_in_minutes: 30
     inputs: [test/sql-server-cdc]
-    parallelism: 3
+    parallelism: 4
     plugins:
       - ./ci/plugins/mzcompose:
           composition: sql-server-cdc
